backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==1.0.3
pytest==7.4.0
requests==2.31.0 